Refactored end_to_end.cpp and moved JIT config specific code to a new file.
