+++ tests/unit/test.vcxproj
@@ -218,6 +218,7 @@
     <ClCompile Include="..\..\libs\runtime\unit\platform_unit_test.cpp" />
     <ClCompile Include="..\..\libs\thunk\mock\mock.cpp" />
     <ClCompile Include="..\end_to_end\end_to_end.cpp" />
+    <ClCompile Include="..\end_to_end\end_to_end_jit.cpp" />
     <ClCompile Include="..\end_to_end\netsh_test.cpp" />
     <ClCompile Include="..\end_to_end\test_helper.cpp" />
     <ClCompile Include="ebpfcore_test.cpp" />
@@ -228,6 +229,7 @@
   <ItemGroup>
     <ClInclude Include="..\..\libs\execution_context\unit\execution_context_unit_test_jit.h" />
     <ClInclude Include="..\..\libs\thunk\mock\mock.h" />
+    <ClInclude Include="..\end_to_end\end_to_end_jit.h" />
     <ClInclude Include="..\end_to_end\helpers.h" />
     <ClInclude Include="..\end_to_end\test_helper.hpp" />
     <ClInclude Include="libbpf_test_jit.h" />
